All the CL/MPI functionality is now transparent.
The main.c is acting as an external developer that wants to use
libhipeless through the header file. He can specify wether to use MPI
or not and using CPU or GPU to execute the OpenCL function with the
use of a flag, he doesn't need to setup OpenCL or MPI in any way.

MPI processes will spawn in the mpihelper function that will direct
them in the correct way.

Note that this branch is not over as the number of mpi processes and
the common variables are still hardcoded.